test(search): Add and consolidate all sealing tests into test_sealing.py
Sealing tests were split across tests.py (OpinionCluster seal admin
views/actions, including the pre-existing coverage from #7799) and
tests_es_recap.py (RECAPDocument seal_documents, including its ES side
effects), each mixed in with unrelated tests. Consolidate them into a
dedicated test_sealing.py, and add coverage for the new
delete_cluster_files behavior:

- OpinionClusterSealClustersActionTest: extracted from
  AdminActionsTest.test_seal_cluster_action (AdminActionsTest keeps
  only its get_search_results tests, trimmed of now-unused fixtures).
- OpinionClusterAdminSealViewTest, OpinionClusterSealDocketBlockersTest,
  OpinionClusterSealActionPermissionTest: moved verbatim from tests.py.
- OpinionClusterSealFileCleanupTest: new, covers the full file cleanup
  (opinion, cluster, and docket files) with a matching CDN invalidation,
  a no-op when nothing to clean up, and leaving the docket's file alone
  when the docket survives sealing.
- RECAPDocumentSealActionESTest: extracted from
  RECAPIndexingTest.test_seal_documents_action in tests_es_recap.py,
  with its own minimal ES setup.

test_re_index_recap_documents_sealed stays in tests_es_recap.py since
its primary subject is the re-index command, not the seal action.
